vantagepro2: Move vantagepro2 agent into package
