+++ tests/agents/test_vantage_pro2_agent.py
@@ -1,3 +1,1 @@
-import sys
-sys.path.insert(0, '../agents/vantagePro2_agent/')
-from vantage_pro2_agent import VantagePro2Agent  # noqa: F401
+from socs.agents.vantagepro2.agent import VantagePro2Agent  # noqa: F401
